# src/main/resources/v1/data-catalog-openapi.yaml
openapi: 3.0.1
info:
  title: Data Catalog
  description: API for Data Catalog
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: PDLEAMZEUG@pdl.internal.ericsson.com
  license:
    name: COPYRIGHT Ericsson "201"9
    url: https://www.ericsson.com/en/legal
  version: 1.0.0
servers:
- url: //localhost:8080/catalog
...
paths:
...
  /v1/message-status-topic:
    get:
      tags:
      - /v1/message-status-topic
      summary: get all MessageStatusTopic details or filter by QueryParams
      operationId: getAllMessageStatusTopicsUsingGET
      parameters:
      - name: messageBusId
        in: query
        description: messageBusId related with MessageStatusTopic
        schema:
          type: string
      - name: name
        in: query
        description: name of the MessageStatusTopic
        schema:
          type: string
      responses:
        "200":
          description: get MessageStatusTopic details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageStatusTopicDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: If MessageStatusTopic details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
    post:
      tags:
      - /v1/message-status-topic
      summary: create MessageStatusTopic details
      operationId: registerMessageStatusTopicUsingPOST
      requestBody:
        description: MessageStatusTopic
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageStatusTopicDto'
        required: true
      responses:
        "201":
          description: created MessageStatusTopic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageStatusTopicDto'
        "400":
          description: Bad Request in values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: MessageStatusTopic details exists already
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
      x-codegen-request-body-name: MessageStatusTopic model
  /v1/message-status-topic/{id}:
    get:
      tags:
      - /v1/message-status-topic
      summary: get MessageStatusTopic by Id
      operationId: getMessageStatusTopicByIdUsingGET
      parameters:
      - name: id
        in: path
        description: id uniquely identifies MessageStatusTopic
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get MessageStatusTopic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageStatusTopicDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: If MessageStatusTopic details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
...
components:
  schemas:
...
    CatalogResponse:
      title: CatalogResponse
      type: object
      properties:
        message:
          type: string
        path:
          type: string
        timeStamp:
          type: string
          format: date-time
...
    MessageBusDto:
      title: MessageBusDto
      type: object
      properties:
        accessEndpoints:
          type: array
          items:
            type: string
        clusterName:
          type: string
        id:
          type: integer
          format: int32
        messageDataTopicIds:
          type: array
          items:
            type: integer
            format: int32
        messageStatusTopicIds:
          type: array
          items:
            type: integer
            format: int32
        name:
          type: string
        nameSpace:
          type: string
        notificationTopicIds:
          type: array
          items:
            type: integer
            format: int32
...
    MessageStatusTopicDto:
      title: MessageStatusTopicDto
      type: object
      properties:
        encoding:
          type: string
          enum:
          - JSON
          - AVRO
          - PROTOBUF
        id:
          type: integer
          format: int32
        messageBus:
          $ref: '#/components/schemas/MessageBusDto'
        messageBusId:
          type: integer
          format: int32
        messageDataTopicIds:
          type: array
          items:
            type: integer
            format: int32
        name:
          type: string
        specificationReference:
          type: string
...